Fixed failure to run unit-tests on non Linux
