+++ edge/pkg/metamanager/metaserver/server_other.go
@@ -1,0 +1,7 @@
+//go:build !linux
+
+package metaserver
+
+func setupDummyInterface() error {
+	return nil
+}
